// frontend/src/api/generated/models/importOpeningTypeEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Dekopen API
 * Authenticated tenant and engine API boundary.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `AWNING` - AWNING
 * * `DOOR_ENTRY` - DOOR_ENTRY
 * * `FIXED` - FIXED
 * * `SLIDING_2L` - SLIDING_2L
 * * `TILT_TURN_LEFT` - TILT_TURN_LEFT
 * * `TILT_TURN_RIGHT` - TILT_TURN_RIGHT
 * * `TURN_LEFT` - TURN_LEFT
 * * `TURN_RIGHT` - TURN_RIGHT
 */
export type ImportOpeningTypeEnum =
  (typeof ImportOpeningTypeEnum)[keyof typeof ImportOpeningTypeEnum];

export const ImportOpeningTypeEnum = {
  AWNING: "AWNING",
  DOOR_ENTRY: "DOOR_ENTRY",
  FIXED: "FIXED",
  SLIDING_2L: "SLIDING_2L",
  TILT_TURN_LEFT: "TILT_TURN_LEFT",
  TILT_TURN_RIGHT: "TILT_TURN_RIGHT",
  TURN_LEFT: "TURN_LEFT",
  TURN_RIGHT: "TURN_RIGHT",
} as const;